+++ data/src/commonMain/kotlin/org/neotech/app/abysner/data/diveplanning/resources/DivePlanInputResourceV1.kt
@@ -33,7 +33,7 @@ data class DivePlanInputResourceV1(
     @Serializable
     data class ProfileSegmentResource(
         val duration: Int,
-        val depth: Int,
+        val depth: Double,
         val cylinderIdentifier: String
     )
